Use hostname --fqdn instead of querying ipinfo.io for hostname
Querying ipinfo.io was the initial solution but it only works if the
reverse DNS lookup matches the actual routing. This is the case with
physical machines but not with VMs, which have trouble getting the
correct hostname to show up.

A better alternative that this commit changes to is to use the
hostname --fqdn command which gives the same info but working on VMs
as well.